move `kind` to `Finding`
